Cambio de tipo de numero_nivel de entero a numeric en la tabla tnivel_organizacional

+++ sis_organigrama/base/funciones/orga.f_get_uo_departamento.sql
@@ -13,7 +13,7 @@ DECLARE
     v_id_uo					integer;
     v_gerencia			varchar;
     v_id_uo_hijo		integer;
-    v_nivel				integer;
+    v_nivel				numeric;
 BEGIN
   	v_nombre_funcion = 'orga.f_get_uo_departamento';
     if (par_id_uo is not null) then
